Sync GitHub with packaged project (#2)
Synchronize the repository with the normalized stm32h7-pocketmod_dsp3d-avi-infoframe-1080p60-default package.

+++ Core/Inc/dsp3D_LL.h
@@ -79,8 +79,8 @@ void dsp3D_LL_init(void);
  * @param[in]  y      y position
  * @param[in]  color  The color
  */
-void dsp3D_LL_drawPoint(uint32_t x, uint32_t y, uint32_t color);
-void dsp3D_LL_drawPointF(uint16_t x, uint16_t y);
+void dsp3D_LL_drawPoint(int32_t x, int32_t y, uint32_t color);
+void dsp3D_LL_drawPointF(int32_t x, int32_t y);
 
 /**
  * @brief      Clear the screen with the specified color
